src/prompt_manager/requirements.txt
weave
python-dotenv 